Move package pkg/process-agent/app/subcommands => pkg/process-agent/subcommands
