+++ cmd/process-agent/subcommands/status/status.go
