+++ cmd/process-agent/subcommands/status/status_test.go
