+++ cmd/process-agent/subcommands/subcommands.go
@@ -6,14 +6,14 @@
 package subcommands
 
 import (
-	cmdcheck "github.com/DataDog/datadog-agent/cmd/process-agent/app/subcommands/check"
-	cmdconfig "github.com/DataDog/datadog-agent/cmd/process-agent/app/subcommands/config"
-	cmdevents "github.com/DataDog/datadog-agent/cmd/process-agent/app/subcommands/events"
-	cmdstatus "github.com/DataDog/datadog-agent/cmd/process-agent/app/subcommands/status"
-	cmdtaggerlist "github.com/DataDog/datadog-agent/cmd/process-agent/app/subcommands/taggerlist"
-	cmdversion "github.com/DataDog/datadog-agent/cmd/process-agent/app/subcommands/version"
-	cmdworkloadlist "github.com/DataDog/datadog-agent/cmd/process-agent/app/subcommands/workloadlist"
 	"github.com/DataDog/datadog-agent/cmd/process-agent/command"
+	cmdcheck "github.com/DataDog/datadog-agent/cmd/process-agent/subcommands/check"
+	cmdconfig "github.com/DataDog/datadog-agent/cmd/process-agent/subcommands/config"
+	cmdevents "github.com/DataDog/datadog-agent/cmd/process-agent/subcommands/events"
+	cmdstatus "github.com/DataDog/datadog-agent/cmd/process-agent/subcommands/status"
+	cmdtaggerlist "github.com/DataDog/datadog-agent/cmd/process-agent/subcommands/taggerlist"
+	cmdversion "github.com/DataDog/datadog-agent/cmd/process-agent/subcommands/version"
+	cmdworkloadlist "github.com/DataDog/datadog-agent/cmd/process-agent/subcommands/workloadlist"
 )
 
 // ProcessAgentSubcommands returns SubcommandFactories for the subcommands in the Process Agent
